LPS-94642 you're right, it's not a DTO, but it's still so tied to DTOs that I'm going to leave it in there. But I'm renaming it to DataRecordValues (plural) to match DataRecord#dataRecordValues
