+++ modules/apps/data-engine/data-engine-rest-impl/src/main/java/com/liferay/data/engine/rest/internal/dto/v1_0/util/DataRecordValuesUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.data.engine.rest.internal.util.v1_0;
+package com.liferay.data.engine.rest.internal.dto.v1_0.util;
 
 import com.liferay.data.engine.rest.dto.v1_0.DataDefinition;
 import com.liferay.data.engine.rest.dto.v1_0.DataDefinitionField;
@@ -32,7 +32,7 @@ import java.util.stream.StreamSupport;
 /**
  * @author Jeyvison Nascimento
  */
-public class DataRecordValueUtil {
+public class DataRecordValuesUtil {
 
 	public static Map<String, ?> toDataRecordValues(
 			DataDefinition dataDefinition, String json)
