+++ modules/apps/data-engine/data-engine-rest-impl/src/main/java/com/liferay/data/engine/rest/internal/storage/DataRecordExporter.java
@@ -17,7 +17,7 @@ package com.liferay.data.engine.rest.internal.storage;
 import com.liferay.data.engine.rest.dto.v1_0.DataDefinition;
 import com.liferay.data.engine.rest.dto.v1_0.DataRecord;
 import com.liferay.data.engine.rest.internal.dto.v1_0.util.DataDefinitionUtil;
-import com.liferay.data.engine.rest.internal.util.v1_0.DataRecordValueUtil;
+import com.liferay.data.engine.rest.internal.dto.v1_0.util.DataRecordValuesUtil;
 import com.liferay.dynamic.data.lists.model.DDLRecordSet;
 import com.liferay.dynamic.data.lists.service.DDLRecordSetLocalService;
 import com.liferay.petra.string.StringPool;
@@ -69,7 +69,7 @@ public class DataRecordExporter {
 		DataDefinition dataDefinition, DataRecord dataRecord) {
 
 		try {
-			return DataRecordValueUtil.toJSON(
+			return DataRecordValuesUtil.toJSON(
 				dataDefinition, dataRecord.getDataRecordValues());
 		}
 		catch (Exception e) {
